Refactor ProductList
